A better solution for this overflow issue.

+++ otserv/trunk/luascript.cpp
@@ -1006,10 +1006,10 @@ int32_t LuaScriptInterface::popCallback(lua_State *L)
 	return luaL_ref(L, LUA_REGISTRYINDEX);
 }
 
-void LuaScriptInterface::setField(lua_State *L, const char* index, int64_t val)
+void LuaScriptInterface::setField(lua_State *L, const char* index, double val)
 {
 	lua_pushstring(L, index);
-	lua_pushnumber(L, (double)val);
+	lua_pushnumber(L, val);
 	lua_settable(L, -3);
 }
