+++ otserv/trunk/luascript.h
@@ -318,7 +318,7 @@ public:
 
 	static int32_t getField(lua_State *L, const char *key);
 	static uint32_t getFieldU32(lua_State *L, const char *key);
-	static void setField(lua_State *L, const char* index, int64_t val);
+	static void setField(lua_State *L, const char* index, double val);
 	static void setField(lua_State *L, const char* index, const std::string& val);
 	static std::string getFieldString(lua_State *L, const char *key);
 	static void setFieldBool(lua_State *L, const char* index, bool val);
